requirements.txt
discord.py[voice]==1.2.5
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
psycopg2==2.8.3
gunicorn==19.9.0
beautifulsoup4==4.7.1
schedule==0.6.0
pytz==2019.2
mecab==0.996
http://download.pytorch.org/whl/cpu/torch-1.0.1-cp37-cp37m-linux_x86_64.whl
numpy==1.18.1
chainer==7.1.0
matplotlib==3.1.2
pyyaml==5.3
parallel-wavegan==0.2.7